Separated rook movement logic a different file
